prevent players destroying equipped items

+++ src/commands/destroy.js
@@ -9,6 +9,7 @@ class Destroy extends Command
         if(this.arguments[0]) {
             const item = await itemsModel.getForUser(this.arguments[0], character.id);
             if(item) {
+                if(character.getEquippedIDs().indexOf(item.id) > -1) return this.message.channel.send(`**${character.name}** you can't destroy equipped items..`);
                 await itemsModel.delete(item.id);
                 return this.message.channel.send(`**${character.name}** destroyed a level ${item.level} ${item.name}`);
             }
